added team cost graph
